requirements.txt
pandas
numpy
scikit-learn
setuptools
google-cloud-storage
pyyaml
tensorflow
comet-ml
dvc
dvc-gs
flask
joblib

